refactor WMATA into its own client
