+++ src/types/trmnl/types.ts
@@ -1,0 +1,6 @@
+export type TrmnlMeta = {
+  user?: { name?: string; time_zone_iana?: string; utc_offset?: number }
+  device?: { friendly_id?: string; percent_charged?: number; wifi_strength?: number; height?: number; width?: number }
+  system?: { timestamp_utc?: number }
+  plugin_settings?: { instance_name?: string }
+}
